Update dev compiler option...
I just installed this and couldn't get dev mode on in order to use the Svelte extension in Chrome. I was using the `{dev: true}` option, but it wouldn't go. Once I changed it to 
```compilerOptions: {
            dev: true,
        },``` 
per the svelte-loader docs, it worked. Maybe I'm missing something, but I thought I'd just share this here.